task2 redone
